add first half denseformer (fix for fp16)

+++ train_gpt.py
@@ -28,7 +28,7 @@ class InPlaceSetSlice(torch.autograd.Function):
     def forward(ctx, full_tensor, last_slice, x_idx, x_val):
         full_tensor[x_idx] = x_val
         ctx.x_idx = x_idx
-        ret = torch.Tensor().to(full_tensor.device)
+        ret = torch.Tensor().to(full_tensor.device, dtype=full_tensor.dtype)
         ret.set_(full_tensor[:x_idx + 1])
         return ret
